Updated, Web Version

+++ src/java/edu/csu2017sp314/DTR14/tripco/WebServer.java
@@ -14,11 +14,9 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.StringReader;
+import java.net.URL;
 import java.nio.ByteBuffer;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import javax.json.Json;
 
 import javax.websocket.CloseReason;
@@ -30,22 +28,19 @@ import javax.websocket.OnOpen;
 import javax.websocket.Session;
 import javax.websocket.server.ServerEndpoint;
 import javax.json.JsonObject;
-import javax.json.JsonObjectBuilder;
 import javax.json.JsonReader;
-import javax.json.JsonValue;
 import javax.websocket.RemoteEndpoint;
-
-@ServerEndpoint("/Server")
+@ServerEndpoint("/WebServer")
 public class WebServer {
     private static File uploadedFile = null;
     private static String fileName = null;
     private static FileOutputStream fos = null;
-    private final static String filePath="./../../../../../TripCoOnline/src/main/webapp/webTemFile/";
+    private final static String workdir = WebServer.class.getProtectionDomain().getCodeSource().getLocation().getPath();
+    private final static String filePath= workdir.substring(0, workdir.indexOf("WEB-INF/"));   
     private final static HashMap<Session, String> maps = new HashMap<Session, String>();
     
     @OnOpen
     public void open(Session session, EndpointConfig conf) {
-        new File(filePath + session.getId()).mkdir();
         System.out.println("Server: DTR-14 Server open successfully");
     }
     
@@ -69,7 +64,7 @@ public class WebServer {
     private void processFile(Session session, String fileName) {   
         System.out.println("Server: " + fileName);
         if (fileName.contains("csv")) maps.put(session, fileName.substring(0, fileName.indexOf('.')));
-        uploadedFile = new File(filePath + session.getId() + "/" + fileName);
+        uploadedFile = new File(filePath + session.getId() + "-" + fileName);
         try {
             fos = new FileOutputStream(uploadedFile);
         } catch (FileNotFoundException e) {     
@@ -80,7 +75,7 @@ public class WebServer {
     }
         
     @OnMessage
-    public void message(Session session, String msg) {
+    public void message(Session session, String msg) throws Exception {
         System.out.println("msg = " + msg);
         JsonReader reader = Json.createReader(new StringReader(msg));
         JsonObject jso = reader.readObject();
@@ -106,12 +101,12 @@ public class WebServer {
 
     @OnClose
     public void close(Session session, CloseReason reason) {
-        File filedir = new File(filePath + session.getId());
+        File filedir = new File(filePath);
         String[] files = filedir.list();
         for(String s: files){
             File currentFile = new File(filedir.getPath(),s);
-            if (s.contains("csv")) maps.remove(session);
-            currentFile.delete();
+            if (s.contains("csv") && s.contains(session.getId())) maps.remove(session);
+            if (currentFile.getName().contains(session.getId())) currentFile.delete();
         }
         filedir.delete();
         
@@ -124,21 +119,22 @@ public class WebServer {
     }
     
     private void listFile(Session session){
-        File filedir = new File(filePath + session.getId());
+        File filedir = new File(filePath);
         String[] files = filedir.list();
         String fnames = "";
         for(String s: files){   
-            fnames += s + "  ";
+            if(s.contains(session.getId()))
+                fnames += s.substring(s.lastIndexOf('-')+1, s.length()) + " ";
         }
         if(fnames.equals("")) fnames += "No Files In Storage";
         sendBackInformation(session, fnames);
     }
     
     private void deleteFile(Session session, String ext){
-        File filedir = new File(filePath + session.getId());
+        File filedir = new File(filePath);
         String[] files = filedir.list();
         for(String s: files){   
-            if(s.toUpperCase().contains(ext)){
+            if(s.toUpperCase().contains(session.getId() + ext)){
                 File currentFile = new File(filedir.getPath(),s);
                 if (s.contains("csv")) maps.remove(session);
                 currentFile.delete();
@@ -148,11 +144,11 @@ public class WebServer {
     }
     
     private String selectFile(Session session, String ext){
-        File filedir = new File(filePath + session.getId());
+        File filedir = new File(filePath);
         String[] files = filedir.list();
         for(String s: files){   
-            if(s.toUpperCase().contains(ext))
-                return filedir.getPath();
+            if(s.contains(session.getId()) && s.toUpperCase().contains(ext))
+                return filedir+"/"+s;
         }
         return null;
     }
@@ -170,25 +166,29 @@ public class WebServer {
     }
     
     private boolean checkValid(Session session){
-        File filedir = new File(filePath + session.getId());
+        File filedir = new File(filePath);
         String[] files = filedir.list();
         int[] couts = {1,1,1};
         for(String s: files){   
-            if(s.toUpperCase().contains("CSV")) couts[0]--;
-            if(s.toUpperCase().contains("SVG")) couts[1]--;
-            if(s.toUpperCase().contains("XML")) couts[2]--;
+            System.out.println(s);
+            if(s.toUpperCase().contains("CSV") && s.contains(session.getId())) couts[0]--;
+            if(s.toUpperCase().contains("SVG") && s.contains(session.getId())) couts[1]--;
+            if(s.toUpperCase().contains("XML") && s.contains(session.getId())) couts[2]--;
         }
         return (couts[0] == 0 ? (couts[1] >= 0 ? (couts[2] >= 0 ? true : false) : true) : false);
     }
     
-    private void startPlanTrip(Session session, String _n, String _m, String _i, String _2, String _3){
+    private void startPlanTrip(Session session, String _n, String _m, String _i, String _2, String _3) throws FileNotFoundException, Exception, Exception{
         
         //TODO:
         String _xml = selectFile(session, "XML");
         String _csv = selectFile(session, "CSV");
         String _svg = selectFile(session, "SVG");
-        boolean[] opts = {false, _i.equals("true"), _m.equals("true"), _n.equals("true"),
-                            _2.equals("true"), _3.equals("true")};
+        System.out.println("Server: xml = " + _xml);
+        System.out.println("Server: csv = " + _csv);
+         System.out.println("Server: svg = " + _svg);       
+        boolean[] opts = {false, _i.equals("true"), _m.equals("true"), 
+                            _n.equals("true"),_2.equals("true"), _3.equals("true")};
         if (_csv == null){
             sendBackInformation(session, "Needs CSV file, please check again");
             return;
@@ -197,9 +197,12 @@ public class WebServer {
             sendBackInformation(session, "No more than one same type file, please check again");
             return;
         }
-        
-//      static ArrayList<File> files;
-        //co.initiate();
+//        public TripCo(String _csv, String _xml, String _svg, String fileroot, 
+//					boolean _gui, boolean _id, boolean _mileage, boolean _name, 
+//						boolean _2opt, boolean _3opt){
+        TripCo co = new TripCo(_xml, _svg, filePath, opts);
+        co.addFile(new File(_csv));
+        co.initiate();
         
         JsonObject json = Json.createObjectBuilder()
                .add("Key", "Result")
